Save layout setting for WidgetReflector. Add menu to select column visibility. Setup WidgetSnapshotVisualizer with clipping zone. Move options out of WidgetReflector. Change WidgetReflector layout. Add filter to WidgetReflector tree. Can now filter only UMG widget. Visualizer key input in DemoMode. Use command `Slate.DemoMode.KeyEvent`or `Slate.DemoMode.MouseEvent`. It is now seperated from the WidgetReflector. Add to HeaderRow an option to select the header that will be generated. WidgetReflector: Add support for GlobalInvalidation. #jira none #rb vincent.gauthier, matt.kuhlenschmidt
#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: patrick.boutot
#ROBOMERGE-SOURCE: CL 14803671 via CL 14804050 via CL 14804085 via CL 14804086
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v760-14685797)

[CL 14804269 by patrick boutot in Dev-Tools-Staging branch]

+++ Engine/Source/Developer/SlateReflector/Private/Widgets/SAtlasVisualizer.cpp
@@ -491,7 +491,7 @@ FText SAtlasVisualizer::GetViewportSizeText() const
 FVector2D SAtlasVisualizer::GetViewportWidgetSize() const
 {
 	const FIntPoint ViewportSize = GetSize();
-	return FVector2D(ViewportSize.X, ViewportSize.Y);
+	return FVector2D((float)ViewportSize.X, (float)ViewportSize.Y);
 }
 
 FText SAtlasVisualizer::GetZoomLevelPercentText() const
